+++ src/Providers.ts
@@ -7,10 +7,13 @@ import { Bucket, BucketProvider } from "./Bucket.ts";
 import { Container, ContainerProvider } from "./Container.ts";
 import * as Credentials from "./Credentials.ts";
 import { Domain, DomainProvider } from "./Domain.ts";
+import { FlexibleIp, FlexibleIpProvider } from "./FlexibleIp.ts";
 import { Namespace, NamespaceProvider } from "./Namespace.ts";
+import { PrivateNic, PrivateNicProvider } from "./PrivateNic.ts";
 import { PrivateNetwork, PrivateNetworkProvider } from "./PrivateNetwork.ts";
 import { RegistryNamespace, RegistryNamespaceProvider } from "./RegistryNamespace.ts";
 import { Secret, SecretProvider } from "./Secret.ts";
+import { SecurityGroup, SecurityGroupProvider } from "./SecurityGroup.ts";
 import { Trigger, TriggerProvider } from "./Trigger.ts";
 import { Vpc, VpcProvider } from "./Vpc.ts";
 import { VpcAcl, VpcAclProvider } from "./VpcAcl.ts";
@@ -37,6 +40,9 @@ export const providers = () =>
       VpcAcl,
       VpcRoute,
       VpcConnector,
+      SecurityGroup,
+      FlexibleIp,
+      PrivateNic,
     ]),
   ).pipe(
     Layer.provide(
@@ -53,6 +59,9 @@ export const providers = () =>
         VpcAclProvider(),
         VpcRouteProvider(),
         VpcConnectorProvider(),
+        SecurityGroupProvider(),
+        FlexibleIpProvider(),
+        PrivateNicProvider(),
       ),
     ),
     Layer.provideMerge(Credentials.fromAuthProvider()),