+++ test/support/test-providers.ts
@@ -8,11 +8,14 @@ import { Bucket, BucketProvider } from "../../src/Bucket.ts";
 import { Container, ContainerProvider } from "../../src/Container.ts";
 import { ScalewayCredentials } from "../../src/Credentials.ts";
 import { Domain, DomainProvider } from "../../src/Domain.ts";
+import { FlexibleIp, FlexibleIpProvider } from "../../src/FlexibleIp.ts";
 import { Namespace, NamespaceProvider } from "../../src/Namespace.ts";
+import { PrivateNic, PrivateNicProvider } from "../../src/PrivateNic.ts";
 import { PrivateNetwork, PrivateNetworkProvider } from "../../src/PrivateNetwork.ts";
 import { Providers } from "../../src/Providers.ts";
 import { RegistryNamespace, RegistryNamespaceProvider } from "../../src/RegistryNamespace.ts";
 import { Secret, SecretProvider } from "../../src/Secret.ts";
+import { SecurityGroup, SecurityGroupProvider } from "../../src/SecurityGroup.ts";
 import { Trigger, TriggerProvider } from "../../src/Trigger.ts";
 import { Vpc, VpcProvider } from "../../src/Vpc.ts";
 import { VpcAcl, VpcAclProvider } from "../../src/VpcAcl.ts";
@@ -46,6 +49,9 @@ export const testProviders = () =>
       VpcAcl,
       VpcRoute,
       VpcConnector,
+      SecurityGroup,
+      FlexibleIp,
+      PrivateNic,
     ]),
   ).pipe(
     Layer.provide(
@@ -62,6 +68,9 @@ export const testProviders = () =>
         VpcAclProvider(),
         VpcRouteProvider(),
         VpcConnectorProvider(),
+        SecurityGroupProvider(),
+        FlexibleIpProvider(),
+        PrivateNicProvider(),
       ),
     ),
     Layer.provideMerge(credentialsLayer),